requirements.txt
flask
python-dotenv
pymupdf
faiss-cpu
sentence-transformers
ollama
flask-login
flask-sqlalchemy
werkzeug